refactor: move evel methods into one module
