+++ tests/rlplg/learning/tabular/test_policyeval.py
@@ -14,12 +14,11 @@ import pytest
 
 from rlplg import core
 from rlplg.learning.opt import schedules
-from rlplg.learning.tabular import policies
-from rlplg.learning.tabular.evaluation import onpolicy
+from rlplg.learning.tabular import policies, policyeval
 from tests import defaults
 
 
-def test_first_visit_monte_carlo_action_values_with_one_episode(
+def test_onpolicy_first_visit_monte_carlo_action_values_with_one_episode(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -33,7 +32,7 @@ def test_first_visit_monte_carlo_action_values_with_one_episode(
 
     """
 
-    results = onpolicy.first_visit_monte_carlo_action_values(
+    results = policyeval.onpolicy_first_visit_monte_carlo_action_values(
         policy=policy,
         environment=environment,
         num_episodes=1,
@@ -52,7 +51,7 @@ def test_first_visit_monte_carlo_action_values_with_one_episode(
     )
 
 
-def test_first_visit_monte_carlo_action_values_with_one_episode_convering_every_action(
+def test_onpolicy_first_visit_monte_carlo_action_values_with_one_episode_convering_every_action(
     environment: gym.Env,
 ):
     """
@@ -75,7 +74,7 @@ def test_first_visit_monte_carlo_action_values_with_one_episode_convering_every_
         actions=[0, 1, 0, 1, 0, 1],
     )
 
-    results = onpolicy.first_visit_monte_carlo_action_values(
+    results = policyeval.onpolicy_first_visit_monte_carlo_action_values(
         policy=stochastic_policy,
         environment=environment,
         num_episodes=1,
@@ -97,7 +96,7 @@ def test_first_visit_monte_carlo_action_values_with_one_episode_convering_every_
     )
 
 
-def test_sarsa_action_values_with_one_episode(
+def test_onpolicy_sarsa_action_values_with_one_episode(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -110,7 +109,7 @@ def test_sarsa_action_values_with_one_episode(
     Q(0,0) = 0 + 0.1 * (-10) = -0.1
     """
 
-    results = onpolicy.sarsa_action_values(
+    results = policyeval.onpolicy_sarsa_action_values(
         policy=policy,
         environment=environment,
         num_episodes=1,
@@ -132,7 +131,7 @@ def test_sarsa_action_values_with_one_episode(
     )
 
 
-def test_sarsa_action_values_with_one_episode_convering_every_action(
+def test_onpolicy_sarsa_action_values_with_one_episode_convering_every_action(
     environment: gym.Env,
 ):
     """
@@ -150,7 +149,7 @@ def test_sarsa_action_values_with_one_episode_convering_every_action(
         actions=[0, 1, 0, 1, 0, 1, 0],
     )
 
-    results = onpolicy.sarsa_action_values(
+    results = policyeval.onpolicy_sarsa_action_values(
         policy=stochastic_policy,
         environment=environment,
         num_episodes=1,
@@ -172,7 +171,7 @@ def test_sarsa_action_values_with_one_episode_convering_every_action(
     )
 
 
-def test_first_visit_monte_carlo_state_values_with_one_episode(
+def test_onpolicy_first_visit_monte_carlo_state_values_with_one_episode(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -183,7 +182,7 @@ def test_first_visit_monte_carlo_state_values_with_one_episode(
     Reversed Returns: 0, -1, -2, -3
     """
 
-    results = onpolicy.first_visit_monte_carlo_state_values(
+    results = policyeval.onpolicy_first_visit_monte_carlo_state_values(
         policy=policy,
         environment=environment,
         num_episodes=1,
@@ -199,7 +198,7 @@ def test_first_visit_monte_carlo_state_values_with_one_episode(
     np.testing.assert_array_equal(values, [-3, -2, -1, 0])
 
 
-def test_first_visit_monte_carlo_state_values_with_two_episodes(
+def test_onpolicy_first_visit_monte_carlo_state_values_with_two_episodes(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -213,7 +212,7 @@ def test_first_visit_monte_carlo_state_values_with_two_episodes(
     Reversed Returns: 0, -1, -2, -3
     """
 
-    results = onpolicy.first_visit_monte_carlo_state_values(
+    results = policyeval.onpolicy_first_visit_monte_carlo_state_values(
         policy=policy,
         environment=environment,
         num_episodes=2,
@@ -233,7 +232,7 @@ def test_first_visit_monte_carlo_state_values_with_two_episodes(
     np.testing.assert_array_equal(values, [-3, -2, -1, 0])
 
 
-def test_one_step_td_state_values_with_one_episode(
+def test_onpolicy_one_step_td_state_values_with_one_episode(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -245,7 +244,7 @@ def test_one_step_td_state_values_with_one_episode(
     Rewards: -1, -1, -1, 0
     """
 
-    results = onpolicy.one_step_td_state_values(
+    results = policyeval.onpolicy_one_step_td_state_values(
         policy=policy,
         environment=environment,
         num_episodes=1,
@@ -264,7 +263,7 @@ def test_one_step_td_state_values_with_one_episode(
     np.testing.assert_allclose(values, [-0.1, -0.1, -0.1, 0])
 
 
-def test_one_step_td_state_values_with_two_episodes(
+def test_onpolicy_one_step_td_state_values_with_two_episodes(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -300,7 +299,7 @@ def test_one_step_td_state_values_with_two_episodes(
     V(2) = -0.19
     """
 
-    results = onpolicy.one_step_td_state_values(
+    results = policyeval.onpolicy_one_step_td_state_values(
         policy=policy,
         environment=environment,
         num_episodes=2,
@@ -323,7 +322,7 @@ def test_one_step_td_state_values_with_two_episodes(
     np.testing.assert_allclose(values, [-0.2, -0.2, -0.19, 0])
 
 
-def test_nstep_td_state_values_with_one_nstep_and_one_episode(
+def test_onpolicy_nstep_td_state_values_with_one_nstep_and_one_episode(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -335,7 +334,7 @@ def test_nstep_td_state_values_with_one_nstep_and_one_episode(
     Rewards: -1, -1, -1, 0
     """
 
-    results = onpolicy.nstep_td_state_values(
+    results = policyeval.onpolicy_nstep_td_state_values(
         policy=policy,
         environment=environment,
         num_episodes=1,
@@ -355,7 +354,7 @@ def test_nstep_td_state_values_with_one_nstep_and_one_episode(
     np.testing.assert_allclose(values, [-0.1, -0.1, -0.1, 0])
 
 
-def test_nstep_td_state_values_with_one_nstep_and_two_episodes(
+def test_onpolicy_nstep_td_state_values_with_one_nstep_and_two_episodes(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -391,7 +390,7 @@ def test_nstep_td_state_values_with_one_nstep_and_two_episodes(
     V(2) = -0.19
     """
 
-    results = onpolicy.nstep_td_state_values(
+    results = policyeval.onpolicy_nstep_td_state_values(
         policy=policy,
         environment=environment,
         num_episodes=2,
@@ -415,7 +414,7 @@ def test_nstep_td_state_values_with_one_nstep_and_two_episodes(
     np.testing.assert_allclose(values, [-0.2, -0.2, -0.19, 0])
 
 
-def test_nstep_td_state_values_with_two_nstep_and_one_episode(
+def test_onpolicy_nstep_td_state_values_with_two_nstep_and_one_episode(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -427,7 +426,7 @@ def test_nstep_td_state_values_with_two_nstep_and_one_episode(
     Rewards: -1, -1, -1, 0
     """
 
-    results = onpolicy.nstep_td_state_values(
+    results = policyeval.onpolicy_nstep_td_state_values(
         policy=policy,
         environment=environment,
         num_episodes=1,
@@ -447,7 +446,7 @@ def test_nstep_td_state_values_with_two_nstep_and_one_episode(
     np.testing.assert_allclose(values, [-0.2, -0.2, -0.1, 0])
 
 
-def test_nstep_td_state_values_with_two_nstep_and_two_episodes(
+def test_onpolicy_nstep_td_state_values_with_two_nstep_and_two_episodes(
     environment: gym.Env,
     policy: core.PyPolicy,
 ):
@@ -487,7 +486,7 @@ def test_nstep_td_state_values_with_two_nstep_and_two_episodes(
     V(2) = -0.19
     """
 
-    results = onpolicy.nstep_td_state_values(
+    results = policyeval.onpolicy_nstep_td_state_values(
         policy=policy,
         environment=environment,
         num_episodes=2,
@@ -511,10 +510,307 @@ def test_nstep_td_state_values_with_two_nstep_and_two_episodes(
     np.testing.assert_allclose(values, [-0.39, -0.38, -0.19, 0.0])
 
 
+def test_offpolicy_monte_carlo_action_values_with_one_episode(
+    environment: gym.Env,
+    policy: core.PyPolicy,
+):
+    """
+    Every (state, action) pair is updated, since
+    the behavior and target policies match on every step.
+
+    Trajectory: (0, 1), (1, 1), (2, 1), (3, 1)
+    Reversed: (3, 1), (2, 1), (1, 1), (0, 1)
+    """
+
+    results = policyeval.offpolicy_monte_carlo_action_values(
+        policy=policy,
+        collect_policy=policy,
+        environment=environment,
+        num_episodes=1,
+        gamma=1.0,
+        policy_probability_fn=policy_prob_fn,
+        collect_policy_probability_fn=collect_policy_prob_fn,
+        state_id_fn=defaults.identity,
+        action_id_fn=defaults.identity,
+        initial_qtable=np.zeros(shape=[4, 2], dtype=np.float32),
+    )
+
+    output = list(results)
+    assert len(output) == 1
+    steps, qtable = next(iter(output))
+    assert steps == 4
+    np.testing.assert_array_equal(qtable, [[0, -3], [0, -2], [0, -1], [0, 0]])
+
+
+def test_offpolicy_monte_carlo_action_values_with_two_episodes(
+    environment: gym.Env,
+    policy: core.PyPolicy,
+):
+    """
+    Every (state, action) pair is updated, since
+    the behavior and target policies match on every step.
+    After one episode, the value for the best actions match.
+
+    Trajectory: (0, 1), (1, 1), (2, 1), (3, 1)
+    Reversed: (3, 1), (2, 1), (1, 1), (0, 1)
+    """
+
+    results = policyeval.offpolicy_monte_carlo_action_values(
+        policy=policy,
+        collect_policy=policy,
+        environment=environment,
+        num_episodes=2,
+        gamma=1.0,
+        policy_probability_fn=policy_prob_fn,
+        collect_policy_probability_fn=collect_policy_prob_fn,
+        state_id_fn=defaults.identity,
+        action_id_fn=defaults.identity,
+        initial_qtable=np.zeros(shape=[4, 2], dtype=np.float32),
+    )
+
+    output = list(results)
+    output_iter = iter(output)
+    assert len(output) == 2
+    steps, qtable = next(output_iter)
+    assert steps == 4
+    np.testing.assert_array_equal(qtable, [[0, -3], [0, -2], [0, -1], [0, 0]])
+    steps, qtable = next(output_iter)
+    assert steps == 4
+    np.testing.assert_array_equal(qtable, [[0, -3], [0, -2], [0, -1], [0, 0]])
+
+
+def test_offpolicy_monte_carlo_action_values_with_one_episode_covering_every_action(
+    environment: gym.Env,
+    policy: core.PyPolicy,
+):
+    """
+    Only the last two state-action pairs are updated because
+    after that, the 2nd to last action mistmatches the policy,
+    turning W into zero - i.e. the importance sample ratio collapses.
+    """
+
+    collect_policy = defaults.RoundRobinActionsPolicy(
+        actions=[0, 1, 0, 1, 0, 1],
+    )
+    results = policyeval.offpolicy_monte_carlo_action_values(
+        policy=policy,
+        collect_policy=collect_policy,
+        environment=environment,
+        num_episodes=1,
+        gamma=1.0,
+        policy_probability_fn=policy_prob_fn,
+        collect_policy_probability_fn=collect_policy_prob_fn,
+        state_id_fn=defaults.identity,
+        action_id_fn=defaults.identity,
+        initial_qtable=np.zeros(shape=[4, 2], dtype=np.float32),
+    )
+
+    output = list(results)
+    assert len(output) == 1
+    steps, qtable = next(iter(output))
+    assert steps == 7
+    # we only learn about the last two (state, action) pairs.
+    np.testing.assert_array_equal(qtable, [[0, 0], [0, 0], [-11, -1], [0, 0]])
+
+
+def test_offpolicy_monte_carlo_action_values_step_with_reward_discount():
+    """
+    G <- 0.9*10 + 100 = 109
+    C(s,a) <- 50 + 2 = 52
+    Q(s,a) <- 20 + 2/52 * (109-20) = 255/13
+    W <- 2 * 0.8 = 1.6
+
+    """
+    mc_update = policyeval.offpolicy_monte_carlo_action_values_step(
+        reward=100, returns=10, cu_sum=50, weight=2.0, value=20, rho=0.8, gamma=0.9
+    )
+
+    np.testing.assert_approx_equal(mc_update.returns, 109)
+    np.testing.assert_approx_equal(mc_update.cu_sum, 52)
+    np.testing.assert_approx_equal(mc_update.value, 609 / 26)
+    np.testing.assert_approx_equal(mc_update.weight, 1.6)
+
+
+def test_offpolicy_nstep_sarsa_action_values_with_one_nstep_and_one_episode(
+    environment: gym.Env,
+    policy: core.PyPolicy,
+):
+    """
+    Each step value updates.
+    Trajectory: (0, 1), (1, 1), (2, 1), (3, 1)
+    """
+
+    results = policyeval.offpolicy_nstep_sarsa_action_values(
+        policy=policy,
+        collect_policy=policy,
+        environment=environment,
+        num_episodes=1,
+        lrs=schedules.LearningRateSchedule(
+            initial_learning_rate=0.1, schedule=constant_learning_rate
+        ),
+        gamma=0.95,
+        nstep=1,
+        policy_probability_fn=policy_prob_fn,
+        collect_policy_probability_fn=collect_policy_prob_fn,
+        state_id_fn=defaults.identity,
+        action_id_fn=defaults.identity,
+        initial_qtable=np.zeros(shape=[4, 2], dtype=np.float32),
+    )
+
+    output = list(results)
+    assert len(output) == 1
+    steps, qtable = next(iter(output))
+    assert steps == 4
+    np.testing.assert_array_almost_equal(
+        qtable, np.array([[0.0, -0.1], [0.0, -0.1], [0.0, -0.1], [0.0, 0.0]])
+    )
+
+
+def test_offpolicy_nstep_sarsa_action_values_with_two_nsteps_and_two_episodes(
+    environment: gym.Env,
+    policy: core.PyPolicy,
+):
+    """
+    Every step is updated after n+2 steps, so the final step isn't updated.
+    """
+    results = policyeval.offpolicy_nstep_sarsa_action_values(
+        policy=policy,
+        collect_policy=policy,
+        environment=environment,
+        num_episodes=2,
+        lrs=schedules.LearningRateSchedule(
+            initial_learning_rate=0.1, schedule=constant_learning_rate
+        ),
+        gamma=0.95,
+        nstep=2,
+        policy_probability_fn=policy_prob_fn,
+        collect_policy_probability_fn=collect_policy_prob_fn,
+        state_id_fn=defaults.identity,
+        action_id_fn=defaults.identity,
+        initial_qtable=np.zeros(shape=[4, 2], dtype=np.float32),
+    )
+
+    output = list(results)
+    output_iter = iter(output)
+    assert len(output) == 2
+    steps, qtable = next(output_iter)
+    assert steps == 4
+    np.testing.assert_array_almost_equal(
+        qtable, np.array([[0.0, -0.195], [0.0, -0.195], [0.0, -0.1], [0.0, 0.0]])
+    )
+    steps, qtable = next(output_iter)
+    assert steps == 4
+    np.testing.assert_array_almost_equal(
+        qtable, np.array([[0.0, -0.379525], [0.0, -0.3705], [0.0, -0.19], [0.0, 0.0]])
+    )
+
+
+def test_offpolicy_nstep_sarsa_action_values_with_one_nstep_and_one_episode_covering_every_action(
+    environment: gym.Env,
+    policy: core.PyPolicy,
+):
+    """
+    Every step preceding a correct step is updated.
+    Every step following a mistep isn't.
+    """
+    collect_policy = defaults.RoundRobinActionsPolicy(
+        actions=[0, 1, 0, 1, 0, 1],
+    )
+
+    results = policyeval.offpolicy_nstep_sarsa_action_values(
+        policy=policy,
+        collect_policy=collect_policy,
+        environment=environment,
+        num_episodes=1,
+        lrs=schedules.LearningRateSchedule(
+            initial_learning_rate=0.1, schedule=constant_learning_rate
+        ),
+        gamma=0.95,
+        nstep=1,
+        policy_probability_fn=policy_prob_fn,
+        collect_policy_probability_fn=collect_policy_prob_fn,
+        state_id_fn=defaults.identity,
+        action_id_fn=defaults.identity,
+        initial_qtable=np.zeros(shape=[4, 2], dtype=np.float32),
+    )
+
+    output = list(results)
+    assert len(output) == 1
+    steps, qtable = next(iter(output))
+    assert steps == 7
+    np.testing.assert_array_almost_equal(
+        qtable, np.array([[-1.0, -0.1], [-1.0, -0.1], [-1.0, -0.1], [0.0, 0.0]])
+    )
+
+
+def test_offpolicy_nstep_sarsa_action_values_with_two_nsteps_and_one_episode_covering_every_action(
+    environment: gym.Env,
+    policy: core.PyPolicy,
+):
+    """
+    No step gets updated.
+    """
+    collect_policy = defaults.RoundRobinActionsPolicy(
+        actions=[0, 1, 0, 1, 0, 1],
+    )
+
+    results = policyeval.offpolicy_nstep_sarsa_action_values(
+        policy=policy,
+        collect_policy=collect_policy,
+        environment=environment,
+        num_episodes=1,
+        lrs=schedules.LearningRateSchedule(
+            initial_learning_rate=0.1, schedule=constant_learning_rate
+        ),
+        gamma=0.95,
+        nstep=2,
+        policy_probability_fn=policy_prob_fn,
+        collect_policy_probability_fn=collect_policy_prob_fn,
+        state_id_fn=defaults.identity,
+        action_id_fn=defaults.identity,
+        initial_qtable=np.zeros(shape=[4, 2], dtype=np.float32),
+    )
+
+    output = list(results)
+    assert len(output) == 1
+    steps, qtable = next(iter(output))
+    assert steps == 7
+    np.testing.assert_array_almost_equal(
+        qtable, np.array([[-2.19, 0.0], [-2.19, 0.0], [-2.19, -0.1], [0.0, 0.0]])
+    )
+
+
+def policy_prob_fn(policy: core.PyPolicy, traj: core.TrajectoryStep) -> float:
+    """The policy we're evaluating is assumed to be greedy w.r.t. Q(s, a).
+    So the best action has probability 1.0, and all the others 0.0.
+    """
+    policy_step = policy.action(observation=traj.observation)
+    return np.where(np.array_equal(policy_step.action, traj.action), 1.0, 0.0).item()  # type: ignore
+
+
+def collect_policy_prob_fn(policy: core.PyPolicy, traj: core.TrajectoryStep) -> float:
+    """The behavior policy is assumed to be fixed over the evaluation window.
+    We log probabilities when choosing actions, so we can just use that information.
+    For a random policy on K arms, log_prob = log(1/K).
+    We just have to return exp(log_prob).
+    """
+    del policy
+    prob: float = np.exp(traj.policy_info["log_probability"])
+    return prob
+
+
 def constant_learning_rate(initial_lr: float, episode: int, step: int):
     del episode
     del step
     return initial_lr
+
+
+@pytest.fixture(scope="function")
+def qtable() -> np.ndarray:
+    """
+    Q-table for optimal actions.
+    """
+    return np.array([[-1, 0], [-1, 0], [-1, 0], [0, 0]], np.float32)
 
 
 @pytest.fixture(scope="function")
@@ -524,14 +820,6 @@ def policy(environment: gym.Env, qtable: np.ndarray) -> core.PyPolicy:
         action_values=qtable,
         emit_log_probability=True,
     )
-
-
-@pytest.fixture(scope="function")
-def qtable() -> np.ndarray:
-    """
-    Q-table for optimal actions.
-    """
-    return np.array([[-1, 0], [-1, 0], [-1, 0], [0, 0]], np.float32)
 
 
 @pytest.fixture(scope="function")
